feature/refactor-easy-pagination (#1226)
